Move map component to defaultMap file and reorganize
